+++ sdk/clientcore/http-netty4/src/main/java/io/clientcore/http/netty4/NettyHttpClient.java
@@ -4,6 +4,7 @@
 package io.clientcore.http.netty4;
 
 import io.clientcore.core.http.client.HttpClient;
+import io.clientcore.core.http.client.HttpProtocolVersion;
 import io.clientcore.core.http.models.HttpHeaderName;
 import io.clientcore.core.http.models.HttpRequest;
 import io.clientcore.core.http.models.Response;
@@ -12,58 +13,53 @@ import io.clientcore.core.instrumentation.logging.ClientLogger;
 import io.clientcore.core.models.CoreException;
 import io.clientcore.core.models.ServerSentResult;
 import io.clientcore.core.models.binarydata.BinaryData;
-import io.clientcore.core.models.binarydata.FileBinaryData;
-import io.clientcore.core.models.binarydata.InputStreamBinaryData;
 import io.clientcore.core.utils.AuthenticateChallenge;
 import io.clientcore.core.utils.CoreUtils;
 import io.clientcore.core.utils.ProgressReporter;
 import io.clientcore.core.utils.ServerSentEventUtils;
 import io.clientcore.http.netty4.implementation.ChannelInitializationProxyHandler;
+import io.clientcore.http.netty4.implementation.Netty4AlpnHandler;
 import io.clientcore.http.netty4.implementation.Netty4ChannelBinaryData;
 import io.clientcore.http.netty4.implementation.Netty4EagerConsumeChannelHandler;
+import io.clientcore.http.netty4.implementation.Netty4HandlerNames;
 import io.clientcore.http.netty4.implementation.Netty4ProgressAndTimeoutHandler;
 import io.clientcore.http.netty4.implementation.Netty4ResponseHandler;
 import io.clientcore.http.netty4.implementation.Netty4SslInitializationHandler;
 import io.clientcore.http.netty4.implementation.ResponseBodyHandling;
 import io.clientcore.http.netty4.implementation.ResponseStateInfo;
-import io.clientcore.http.netty4.implementation.WrappedHttpHeaders;
 import io.netty.bootstrap.Bootstrap;
-import io.netty.buffer.ByteBuf;
 import io.netty.buffer.Unpooled;
 import io.netty.channel.Channel;
-import io.netty.channel.ChannelFuture;
 import io.netty.channel.ChannelFutureListener;
 import io.netty.channel.ChannelInitializer;
 import io.netty.channel.EventLoopGroup;
-import io.netty.handler.codec.http.DefaultFullHttpRequest;
-import io.netty.handler.codec.http.DefaultHttpRequest;
-import io.netty.handler.codec.http.HttpMethod;
-import io.netty.handler.codec.http.HttpVersion;
+import io.netty.handler.codec.http2.Http2SecurityUtil;
 import io.netty.handler.proxy.ProxyHandler;
+import io.netty.handler.ssl.ApplicationProtocolConfig;
+import io.netty.handler.ssl.ApplicationProtocolNames;
 import io.netty.handler.ssl.SslContext;
 import io.netty.handler.ssl.SslContextBuilder;
-import io.netty.handler.stream.ChunkedInput;
-import io.netty.handler.stream.ChunkedNioFile;
-import io.netty.handler.stream.ChunkedStream;
-import io.netty.handler.stream.ChunkedWriteHandler;
+import io.netty.handler.ssl.SslHandler;
+import io.netty.handler.ssl.SslProvider;
+import io.netty.handler.ssl.SupportedCipherSuiteFilter;
 
 import javax.net.ssl.SSLException;
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.net.URI;
-import java.nio.channels.FileChannel;
-import java.nio.file.StandardOpenOption;
 import java.util.List;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.atomic.AtomicReference;
+import java.util.function.Consumer;
 
 import static io.clientcore.core.utils.ServerSentEventUtils.attemptRetry;
 import static io.clientcore.core.utils.ServerSentEventUtils.processTextEventStream;
-import static io.clientcore.http.netty4.implementation.Netty4Utility.PROGRESS_AND_TIMEOUT_HANDLER_NAME;
+import static io.clientcore.http.netty4.implementation.Netty4HandlerNames.HTTP_RESPONSE;
+import static io.clientcore.http.netty4.implementation.Netty4HandlerNames.PROGRESS_AND_TIMEOUT;
 import static io.clientcore.http.netty4.implementation.Netty4Utility.awaitLatch;
 import static io.clientcore.http.netty4.implementation.Netty4Utility.createCodec;
+import static io.clientcore.http.netty4.implementation.Netty4Utility.sendHttp11Request;
 import static io.clientcore.http.netty4.implementation.Netty4Utility.setOrSuppressError;
-import static io.netty.handler.codec.http.DefaultHttpHeadersFactory.trailersFactory;
 
 /**
  * HttpClient implementation using synchronous Netty operations.
@@ -78,18 +74,20 @@ class NettyHttpClient implements HttpClient {
         = "No ServerSentEventListener attached to HttpRequest to handle the text/event-stream response";
 
     private final Bootstrap bootstrap;
-    private final SslContext sslContext;
+    private final Consumer<SslContextBuilder> sslContextModifier;
     private final ChannelInitializationProxyHandler channelInitializationProxyHandler;
     private final AtomicReference<List<AuthenticateChallenge>> proxyChallenges;
     private final long readTimeoutMillis;
     private final long responseTimeoutMillis;
     private final long writeTimeoutMillis;
+    private final HttpProtocolVersion maximumHttpVersion;
 
-    NettyHttpClient(Bootstrap bootstrap, SslContext sslContext,
-        ChannelInitializationProxyHandler channelInitializationProxyHandler, long readTimeoutMillis,
-        long responseTimeoutMillis, long writeTimeoutMillis) {
+    NettyHttpClient(Bootstrap bootstrap, Consumer<SslContextBuilder> sslContextModifier,
+        HttpProtocolVersion maximumHttpVersion, ChannelInitializationProxyHandler channelInitializationProxyHandler,
+        long readTimeoutMillis, long responseTimeoutMillis, long writeTimeoutMillis) {
         this.bootstrap = bootstrap;
-        this.sslContext = sslContext;
+        this.sslContextModifier = sslContextModifier;
+        this.maximumHttpVersion = maximumHttpVersion;
         this.channelInitializationProxyHandler = channelInitializationProxyHandler;
         this.proxyChallenges = new AtomicReference<>();
         this.readTimeoutMillis = readTimeoutMillis;
@@ -121,9 +119,9 @@ class NettyHttpClient implements HttpClient {
         // request basis.
         bootstrap.handler(new ChannelInitializer<Channel>() {
             @Override
-            protected void initChannel(Channel ch) throws SSLException {
+            protected void initChannel(Channel channel) throws SSLException {
                 // Test whether proxying should be applied to this Channel. If so, add it.
-                boolean hasProxy = channelInitializationProxyHandler.test(ch.remoteAddress());
+                boolean hasProxy = channelInitializationProxyHandler.test(channel.remoteAddress());
                 if (hasProxy) {
                     ProxyHandler proxyHandler = channelInitializationProxyHandler.createProxy(proxyChallenges);
                     proxyHandler.connectFuture().addListener(future -> {
@@ -132,27 +130,54 @@ class NettyHttpClient implements HttpClient {
                         }
                     });
 
-                    ch.pipeline().addFirst(proxyHandler);
+                    channel.pipeline().addFirst(Netty4HandlerNames.PROXY, proxyHandler);
                 }
 
                 // Add SSL handling if the request is HTTPS.
                 if (isHttps) {
-                    SslContext ssl = (sslContext != null)
-                        ? sslContext
-                        : SslContextBuilder.forClient().endpointIdentificationAlgorithm("HTTPS").build();
-                    // SSL handling is added last here. This is done as proxying could require SSL handling too.
-                    ch.pipeline().addLast(ssl.newHandler(ch.alloc(), host, port), new Netty4SslInitializationHandler());
-                }
+                    SslContextBuilder sslContextBuilder
+                        = SslContextBuilder.forClient().endpointIdentificationAlgorithm("HTTPS");
+                    if (maximumHttpVersion == HttpProtocolVersion.HTTP_2) {
+                        // If HTTP/2 is the maximum version, we need to ensure that ALPN is enabled.
+                        SslProvider sslProvider = SslContext.defaultClientProvider();
+                        ApplicationProtocolConfig.SelectorFailureBehavior selectorBehavior;
+                        ApplicationProtocolConfig.SelectedListenerFailureBehavior selectedBehavior;
+                        if (sslProvider == SslProvider.JDK) {
+                            selectorBehavior = ApplicationProtocolConfig.SelectorFailureBehavior.FATAL_ALERT;
+                            selectedBehavior = ApplicationProtocolConfig.SelectedListenerFailureBehavior.FATAL_ALERT;
+                        } else {
+                            // Netty OpenSslContext doesn't support FATAL_ALERT, use NO_ADVERTISE and ACCEPT
+                            // instead.
+                            selectorBehavior = ApplicationProtocolConfig.SelectorFailureBehavior.NO_ADVERTISE;
+                            selectedBehavior = ApplicationProtocolConfig.SelectedListenerFailureBehavior.ACCEPT;
+                        }
 
-                // Finally add the HttpClientCodec last as it will need to handle processing request and response
-                // writes and reads for not only the actual request but any proxy or SSL handling.
-                ch.pipeline().addLast(createCodec());
+                        sslContextBuilder.ciphers(Http2SecurityUtil.CIPHERS, SupportedCipherSuiteFilter.INSTANCE)
+                            .applicationProtocolConfig(new ApplicationProtocolConfig(
+                                ApplicationProtocolConfig.Protocol.ALPN, selectorBehavior, selectedBehavior,
+                                ApplicationProtocolNames.HTTP_2, ApplicationProtocolNames.HTTP_1_1));
+                    }
+                    if (sslContextModifier != null) {
+                        // Allow the caller to modify the SslContextBuilder before it is built.
+                        sslContextModifier.accept(sslContextBuilder);
+                    }
+
+                    SslContext ssl = sslContextBuilder.build();
+                    // SSL handling is added last here. This is done as proxying could require SSL handling too.
+                    channel.pipeline().addLast(Netty4HandlerNames.SSL, ssl.newHandler(channel.alloc(), host, port));
+                    channel.pipeline()
+                        .addLast(Netty4HandlerNames.SSL_INITIALIZER, new Netty4SslInitializationHandler());
+
+                    channel.pipeline()
+                        .addLast(Netty4HandlerNames.ALPN,
+                            new Netty4AlpnHandler(request, responseReference, errorReference, latch));
+                }
             }
         });
 
         bootstrap.connect(host, port).addListener((ChannelFutureListener) connectListener -> {
             if (!connectListener.isSuccess()) {
-                LOGGER.atError().setThrowable(connectListener.cause()).log("Failed to send request");
+                LOGGER.atError().setThrowable(connectListener.cause()).log("Failed connection.");
                 errorReference.set(connectListener.cause());
                 connectListener.channel().close();
                 latch.countDown();
@@ -173,13 +198,9 @@ class NettyHttpClient implements HttpClient {
             // effectively be a no-op.
             if (addProgressAndTimeoutHandler) {
                 channel.pipeline()
-                    .addLast(PROGRESS_AND_TIMEOUT_HANDLER_NAME, new Netty4ProgressAndTimeoutHandler(progressReporter,
+                    .addLast(PROGRESS_AND_TIMEOUT, new Netty4ProgressAndTimeoutHandler(progressReporter,
                         writeTimeoutMillis, responseTimeoutMillis, readTimeoutMillis));
             }
-
-            Netty4ResponseHandler responseHandler
-                = new Netty4ResponseHandler(request, responseReference, errorReference, latch);
-            channel.pipeline().addLast(responseHandler);
 
             Throwable earlyError = errorReference.get();
             if (earlyError != null) {
@@ -189,16 +210,47 @@ class NettyHttpClient implements HttpClient {
                 return;
             }
 
-            sendRequest(request, channel, addProgressAndTimeoutHandler, errorReference)
-                .addListener((ChannelFutureListener) sendListener -> {
-                    if (!sendListener.isSuccess()) {
-                        setOrSuppressError(errorReference, sendListener.cause());
-                        sendListener.channel().close();
+            // What I basically want here is the following logic in Netty:
+            // 1. If a proxy exists, it should be added first. When the connection is activated, we should connect
+            //    to the proxy (with or without SSL). If there is no proxy, skip this step.
+            // 2. Once step 1 is complete, we should wait until the SSL handshake is complete (if applicable).
+            //    If SSL isn't being used, skip this step.
+            // 3. Once step 2 is complete, we should send the request.
+            //
+            // None of the steps should block the event loop, so we need to use listeners to ensure that the next
+            // step is only executed once the previous step is complete.
+            SslHandler sslHandler = channel.pipeline().get(SslHandler.class);
+            if (sslHandler != null) {
+                // If the SslHandler is present, trigger the SSL handshake to complete before sending the request.
+                sslHandler.handshakeFuture().addListener(handshakeListener -> {
+                    if (!handshakeListener.isSuccess()) {
+                        LOGGER.atError().setThrowable(handshakeListener.cause()).log("Failed SSL handshake.");
+                        errorReference.set(handshakeListener.cause());
                         latch.countDown();
-                    } else {
-                        sendListener.channel().read();
                     }
                 });
+                channel.write(Unpooled.EMPTY_BUFFER);
+            } else {
+                // If there isn't an SslHandler, we can send the request immediately.
+                // Add the HTTP/1.1 codec, as we only support HTTP/2 when using SSL ALPN.
+                Netty4ResponseHandler responseHandler
+                    = new Netty4ResponseHandler(request, responseReference, errorReference, latch);
+                channel.pipeline().addLast(HTTP_RESPONSE, responseHandler);
+
+                String addBefore = addProgressAndTimeoutHandler ? PROGRESS_AND_TIMEOUT : HTTP_RESPONSE;
+                channel.pipeline().addBefore(addBefore, Netty4HandlerNames.HTTP_CODEC, createCodec());
+
+                sendHttp11Request(request, channel, errorReference)
+                    .addListener((ChannelFutureListener) sendListener -> {
+                        if (!sendListener.isSuccess()) {
+                            setOrSuppressError(errorReference, sendListener.cause());
+                            sendListener.channel().close();
+                            latch.countDown();
+                        } else {
+                            sendListener.channel().read();
+                        }
+                    });
+            }
         });
 
         awaitLatch(latch);
@@ -231,7 +283,7 @@ class NettyHttpClient implements HttpClient {
                 // We're ignoring the response content.
                 CountDownLatch drainLatch = new CountDownLatch(1);
                 channel.pipeline().addLast(new Netty4EagerConsumeChannelHandler(drainLatch, ignored -> {
-                }));
+                }, info.isHttp2()));
                 channel.config().setAutoRead(true);
                 awaitLatch(drainLatch);
             } else if (bodyHandling == ResponseBodyHandling.STREAM) {
@@ -249,7 +301,7 @@ class NettyHttpClient implements HttpClient {
                     }
                 }
 
-                body = new Netty4ChannelBinaryData(info.getEagerContent(), channel, length);
+                body = new Netty4ChannelBinaryData(info.getEagerContent(), channel, length, info.isHttp2());
             } else {
                 // All cases otherwise assume BUFFER.
                 CountDownLatch drainLatch = new CountDownLatch(1);
@@ -259,7 +311,7 @@ class NettyHttpClient implements HttpClient {
                     } catch (IOException ex) {
                         throw LOGGER.throwableAtError().log(ex, CoreException::from);
                     }
-                }));
+                }, info.isHttp2()));
                 channel.config().setAutoRead(true);
                 awaitLatch(drainLatch);
 
@@ -302,78 +354,6 @@ class NettyHttpClient implements HttpClient {
         return response;
     }
 
-    private ChannelFuture sendRequest(HttpRequest request, Channel channel, boolean progressAndTimeoutHandlerAdded,
-        AtomicReference<Throwable> errorReference) {
-        HttpMethod nettyMethod = HttpMethod.valueOf(request.getHttpMethod().toString());
-        String uri = request.getUri().toString();
-        WrappedHttpHeaders wrappedHttpHeaders = new WrappedHttpHeaders(request.getHeaders());
-
-        // TODO (alzimmer): This will mutate the underlying ClientCore HttpHeaders. Will need to think about this design
-        //  more once it's closer to completion.
-        wrappedHttpHeaders.getCoreHeaders().set(HttpHeaderName.HOST, request.getUri().getHost());
-
-        BinaryData requestBody = request.getBody();
-        if (requestBody instanceof FileBinaryData) {
-            FileBinaryData fileBinaryData = (FileBinaryData) requestBody;
-            try {
-                return sendChunked(channel,
-                    new ChunkedNioFile(FileChannel.open(fileBinaryData.getFile(), StandardOpenOption.READ),
-                        fileBinaryData.getPosition(), fileBinaryData.getLength(), 8192),
-                    new DefaultHttpRequest(HttpVersion.HTTP_1_1, nettyMethod, uri, wrappedHttpHeaders),
-                    progressAndTimeoutHandlerAdded, errorReference);
-            } catch (IOException ex) {
-                return channel.newFailedFuture(ex);
-            }
-        } else if (requestBody instanceof InputStreamBinaryData) {
-            return sendChunked(channel, new ChunkedStream(requestBody.toStream()),
-                new DefaultHttpRequest(HttpVersion.HTTP_1_1, nettyMethod, uri, wrappedHttpHeaders),
-                progressAndTimeoutHandlerAdded, errorReference);
-        } else {
-            ByteBuf body = Unpooled.EMPTY_BUFFER;
-            if (requestBody != null && requestBody != BinaryData.empty()) {
-                // Longer term, see if there is a way to have BinaryData act as the ByteBuf body to further eliminate
-                // copying of byte[]s.
-                body = Unpooled.wrappedBuffer(requestBody.toBytes());
-            }
-            if (body.readableBytes() > 0) {
-                // TODO (alzimmer): Should we be setting Content-Length here again? Shouldn't this be handled externally
-                //  by the creator of the HttpRequest?
-                wrappedHttpHeaders.getCoreHeaders()
-                    .set(HttpHeaderName.CONTENT_LENGTH, String.valueOf(body.readableBytes()));
-            }
-
-            Throwable error = errorReference.get();
-            if (error != null) {
-                return channel.newFailedFuture(error);
-            }
-
-            return channel.writeAndFlush(new DefaultFullHttpRequest(HttpVersion.HTTP_1_1, nettyMethod, uri, body,
-                wrappedHttpHeaders, trailersFactory().newHeaders()));
-        }
-    }
-
-    private <T> ChannelFuture sendChunked(Channel channel, ChunkedInput<T> chunkedInput,
-        io.netty.handler.codec.http.HttpRequest initialLineAndHeaders, boolean progressAndTimeoutHandlerAdded,
-        AtomicReference<Throwable> errorReference) {
-        if (channel.pipeline().get(ChunkedWriteHandler.class) == null) {
-            // Add the ChunkedWriteHandler which will handle sending the chunkedInput.
-            ChunkedWriteHandler chunkedWriteHandler = new ChunkedWriteHandler();
-            if (progressAndTimeoutHandlerAdded) {
-                channel.pipeline().addBefore(PROGRESS_AND_TIMEOUT_HANDLER_NAME, null, chunkedWriteHandler);
-            } else {
-                channel.pipeline().addLast(chunkedWriteHandler);
-            }
-        }
-
-        Throwable error = errorReference.get();
-        if (error != null) {
-            return channel.newFailedFuture(error);
-        }
-
-        channel.write(initialLineAndHeaders);
-        return channel.writeAndFlush(chunkedInput);
-    }
-
     public void close() {
         EventLoopGroup group = bootstrap.config().group();
         if (group != null) {
@@ -386,4 +366,5 @@ class NettyHttpClient implements HttpClient {
             ? BinaryData.fromString(String.join("\n", serverSentResult.getData()))
             : BinaryData.empty();
     }
+
 }