+++ sdk/clientcore/http-netty4/src/main/java/io/clientcore/http/netty4/implementation/Netty4InitiateOneReadHandler.java
@@ -2,15 +2,18 @@
 // Licensed under the MIT License.
 package io.clientcore.http.netty4.implementation;
 
+import io.clientcore.core.utils.IOExceptionCheckedConsumer;
 import io.netty.buffer.ByteBuf;
 import io.netty.buffer.ByteBufHolder;
 import io.netty.channel.ChannelHandlerContext;
 import io.netty.channel.ChannelInboundHandler;
 import io.netty.channel.ChannelInboundHandlerAdapter;
 import io.netty.handler.codec.http.LastHttpContent;
+import io.netty.handler.codec.http2.Http2DataFrame;
+import io.netty.util.ReferenceCountUtil;
 
+import java.io.IOException;
 import java.util.concurrent.CountDownLatch;
-import java.util.function.Consumer;
 
 /**
  * {@link ChannelInboundHandler} that initiates one read request any time data is needed. Even though it is a single
@@ -18,7 +21,8 @@ import java.util.function.Consumer;
  * this handler need to support multiple channelRead, if this isn't done data may be lost.
  */
 public final class Netty4InitiateOneReadHandler extends ChannelInboundHandlerAdapter {
-    private final Consumer<ByteBuf> byteBufConsumer;
+    private final IOExceptionCheckedConsumer<ByteBuf> byteBufConsumer;
+    private final boolean isHttp2;
 
     private CountDownLatch latch;
 
@@ -34,10 +38,13 @@ public final class Netty4InitiateOneReadHandler extends ChannelInboundHandlerAda
      *
      * @param latch The latch to count down when the channel read completes.
      * @param byteBufConsumer The consumer to process the {@link ByteBuf ByteBufs} as they are read.
+     * @param isHttp2 Flag indicating whether the handler is used for HTTP/2 or not.
      */
-    public Netty4InitiateOneReadHandler(CountDownLatch latch, Consumer<ByteBuf> byteBufConsumer) {
+    public Netty4InitiateOneReadHandler(CountDownLatch latch, IOExceptionCheckedConsumer<ByteBuf> byteBufConsumer,
+        boolean isHttp2) {
         this.latch = latch;
         this.byteBufConsumer = byteBufConsumer;
+        this.isHttp2 = isHttp2;
     }
 
     /**
@@ -67,10 +74,20 @@ public final class Netty4InitiateOneReadHandler extends ChannelInboundHandlerAda
         }
 
         if (buf != null && buf.isReadable()) {
-            byteBufConsumer.accept(buf);
+            try {
+                byteBufConsumer.accept(buf);
+            } catch (IOException | RuntimeException ex) {
+                ReferenceCountUtil.release(buf);
+                ctx.close();
+                return;
+            }
         }
 
-        lastRead = msg instanceof LastHttpContent;
+        if (isHttp2) {
+            lastRead = msg instanceof Http2DataFrame && ((Http2DataFrame) msg).isEndStream();
+        } else {
+            lastRead = msg instanceof LastHttpContent;
+        }
         ctx.fireChannelRead(msg);
     }