+++ sdk/eventgrid/azure-messaging-eventgrid/src/main/java/com/azure/messaging/eventgrid/systemevents/IotHubDeviceTelemetryEventData.java
@@ -13,8 +13,12 @@ import java.util.Map;
 
 /**
  * Event data for Microsoft.Devices.DeviceTelemetry event.
+ * 
+ * @deprecated This class is deprecated and may be removed in future releases. System events are now available in the
+ * azure-messaging-eventgrid-systemevents package.
  */
 @Fluent
+@Deprecated
 public final class IotHubDeviceTelemetryEventData extends DeviceTelemetryEventProperties {
 
     /**